Streaming version of derivative computation.
Completely untested.
Since sums for Z_context are also computed in the loop, some terms are now computed without normalization then normalized afterwards.
The structure of tune_derivatives is now less logical (no gradient then Hessian) due to the need to sum over extensions in one loop.

+++ lm/interpolate/tune_matrix.hh
@@ -8,8 +8,8 @@
 
 namespace lm { namespace interpolate {
  
-typedef Eigen::MatrixXd Matrix;
-typedef Eigen::VectorXd Vector;
+typedef Eigen::MatrixXf Matrix;
+typedef Eigen::VectorXf Vector;
 
 typedef Matrix::Scalar Accum;
